+++ lib/core/errors/app_failure.dart
@@ -88,6 +88,7 @@ enum ValidationErrorKind {
   specialChars,
   duplicateName,
   renamePartialFailure,
+  moveIntoSelf,
 }
 
 class ValidationFailure extends AppFailure {